+++ packages/imperative/src/config/__tests__/ProfileInfo.TeamConfig.unit.test.ts
@@ -1047,6 +1047,31 @@ describe("TeamConfig ProfileInfo tests", () => {
             expect(updateKnownPropertySpy).toHaveBeenCalledWith({ ...profileOptions, mergedArgs, osLocInfo });
         });
 
+        it("should succeed forceUpdating a property even if the property doesn't exist", async () => {
+            const profInfo = createNewProfInfo(teamProjDir);
+            await profInfo.readProfilesFromDisk();
+            const storeSpy = jest.spyOn(ConfigAutoStore, "_storeSessCfgProps").mockImplementation(jest.fn());
+            const profileOptions: IProfInfoUpdatePropOpts = {
+                profileName: "LPAR4",
+                profileType: "dummy",
+                property: "DOES_NOT_EXIST",
+                value: true,
+                forceUpdate: true
+            };
+            let caughtError;
+            try {
+                await profInfo.updateProperty(profileOptions);
+            } catch (error) {
+                caughtError = error;
+            }
+            expect(caughtError).toBeUndefined();
+            expect(storeSpy).toHaveBeenCalledWith({
+                config: profInfo.getTeamConfig(), profileName: "LPAR4", profileType: "dummy",
+                defaultBaseProfileName: "base_glob",
+                propsToStore: [ "DOES_NOT_EXIST" ], sessCfg: { "DOES_NOT_EXIST": true }, setSecure : undefined,
+            });
+        });
+
         it("should attempt to store session config properties without adding profile types to the loadedConfig", async () => {
             const profInfo = createNewProfInfo(teamProjDir);
             await profInfo.readProfilesFromDisk();
@@ -1468,7 +1493,7 @@ describe("TeamConfig ProfileInfo tests", () => {
         });
 
         describe("addProfileTypeToSchema", () => {
-            const expectAddToSchemaTester = async (testCase: { schema: any; previousVersion?: string; version?: string }, expected: {
+            const expectAddToSchemaTester = async (testCase: { schema: any; previousVersion?: string }, expected: {
                 extendersJson: IExtendersJsonOpts,
                 res: {
                     success: boolean;
@@ -1531,7 +1556,7 @@ describe("TeamConfig ProfileInfo tests", () => {
 
             it("only updates a profile type in the schema if the version is newer", async () => {
                 expectAddToSchemaTester(
-                    { previousVersion: "1.0.0", schema: { title: "Mock Schema" } as any, version: "2.0.0" },
+                    { previousVersion: "1.0.0", schema: { title: "Mock Schema", version: "2.0.0" } as any },
                     {
                         extendersJson: {
                             profileTypes: {
@@ -1551,7 +1576,7 @@ describe("TeamConfig ProfileInfo tests", () => {
 
             it("does not update a profile type in the schema if the version is older", async () => {
                 expectAddToSchemaTester(
-                    { previousVersion: "2.0.0", schema: { title: "Mock Schema" } as any, version: "1.0.0" },
+                    { previousVersion: "2.0.0", schema: { title: "Mock Schema", version: "1.0.0" } as any },
                     {
                         extendersJson: {
                             profileTypes: {
@@ -1571,7 +1596,7 @@ describe("TeamConfig ProfileInfo tests", () => {
 
             it("updates a profile type in the schema - version provided, no previous schema version", async () => {
                 expectAddToSchemaTester(
-                    { previousVersion: "none", schema: { title: "Mock Schema" } as any, version: "1.0.0" },
+                    { previousVersion: "none", schema: { title: "Mock Schema", version: "1.0.0" } as any },
                     {
                         extendersJson: {
                             profileTypes: {
@@ -1591,7 +1616,7 @@ describe("TeamConfig ProfileInfo tests", () => {
 
             it("does not update the schema if schema version is invalid", async () => {
                 expectAddToSchemaTester(
-                    { previousVersion: "none", schema: { title: "Mock Schema" } as any, version: "1.0.0" },
+                    { previousVersion: "none", schema: { title: "Mock Schema", version: "1.0.0" } as any },
                     {
                         extendersJson: {
                             profileTypes: {
@@ -1610,7 +1635,6 @@ describe("TeamConfig ProfileInfo tests", () => {
             });
         });
         describe("buildSchema", () => {
-            // TODO
             it("builds a schema with the default types", async () => {
                 const profInfo = createNewProfInfo(teamProjDir);
                 await profInfo.readProfilesFromDisk({ homeDir: teamHomeProjDir });